Lua: evaluate return code for image read callback
On custom design handler, a Lua callback can be set when an image should
be read. The lua interface just checks that the Lua callback can be
executed, but it does not check if the callback returns with an error
code, and SWUpdate cannot detect this. Check return code in istream_read_callback
instead of just returning success.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>