swupdate_async_start: fix race condition
There is a race condition when swupdate-client installs a swu image, leading to
an error exit code of the client, while the daemon successfully installs the
image. Other programs using swupdate_async_start may be impacted.

It is seen on small images (eg: 20480 bytes) on some machines:
- swupdate-client:
     * sends REQ_INSTALL to the daemon
     * ... followed by the payload (the image), which is buffered by the OS
     * and then sends GET_STATUS to the daemon whereas the daemon has not even
       started receiving the payload
- swupdate daemon:
     * receives REQ_INSTALL and sends ACK
     * (1) receives GET_STATUS and answers with IDLE
     * (2) notifies another thread that sets the status to RUN and proceeds with
       the installation
- swupdate-client:
     * receives the status IDLE
     * considers this as a failure, and exits with code 1 (error)
- swupdate daemon:
     * continues normally and the installation succeeds
       (status: IDLE --> RUN --> SUCCESS)

In this patch we do not rely on GET_STATUS any longer but rather use the
"progress" interface to be notified of the end and result of the installation.

We still use GET_STATUS though, to provide compatibility with programs that use
the status_func callback (for swupdate-client, the messages are printed to
stdout).

Signed-off-by: Frederic Hoerni <fhoerni@witekio.com>