BUG: Parser: solve incompatibility between parsers
It is possible to switch between json and libconfig, but JSON allows to
scale down an array to a scalra if it has just one element, and this
breaks the parser looking for array in case of libconfig that is strict
typed. Solve this allowing this singularity in case of libconfig.
Note that this issue causes SWUpdate to abort in json-c library
because a function operating on a JSON array is called - by
checking the type, this is solved, too.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>