util: introduce generic mount helpers
Functions asking to mount / umount filesystems must prepare the same
paths for the mountpoints. Factorize this into two new funtions:

	swupdate_temporary_mount()
	swupdate_temporary_umount()

If the mopunt part isn successful, it returns a generated path (not
fixed anymore) as mountpoint. This should be passed to the corresponding
umount(). that becomes responsible to free memory.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>