progress: acknowledge client registration
The client that connected to the progress socket had no clear confirmation that
the daemon was ready to send progress events.

This patch makes the daemon send an acknowledgement when the subscription is
duly registered.

On the client side, the function progress_ipc_connect() waits for the
acknowledgement before returning.

Signed-off-by: Frederic Hoerni <fhoerni@witekio.com>