swupdate_async_start: fix early termination blocking in client
It could happen on the client side, that the termination status of the
installation was missed, leading to the client waiting forever.

This modification connects the progress socket before sending the image
payload, to be sure not to miss the termination status (SUCCESS or FAILURE).

We also need to consume the events during the image transfer, so that the pipe
does not get full and block the daemon.

Signed-off-by: Frederic Hoerni <fhoerni@witekio.com>