btrfs: introduce support for snapshot with btrfs receive
Add a handler that supports snapshots created by "btrfs send" as
artifact stored into the SWU. The handler needs that btrfs utility is
installed on the system, and reports a runtime error if it is not found.
It is implemented as image handler, and the image is streamed to a
forked process that will run "btrfs receive" in background.

To realize this, a generic way to stream an artifact to an external
command is implemented. This allows to push the stream from the
copyimage() to an external command that accept the data from stdin, like
"btrfs receive" is doing. To extend this, a generic "executor" handler
is provided, too, that allows to set the external command as one of its
property.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>