Drop bootscripts from general structure
Bootloader environment can be installed as "image" with a handler of
type bootloader. There is no reason to collect them in a separate list,
and putting them in the bootenv section does not work.

A list of variables in a file is supported with the "bootloader" handler
if the artifact is in the images section. There is no reason to have it
in the "bootenv" section that is thought to set individual
variable.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>