ci: Call pkg-config to find lua5.2.pc path
setup.sh uses a hardcoded path with the target machine name to create
the lua.pc symbolic link, because of that it is not possible to build
on hosts using different paths (example: ARM architectures).

But, pkg-config store the path for the '.pc' file in the variable
'pcfiledir'. Thus, once lua gets installed, it is possible to detect
its path with

$ pkg-config --variable=pcfiledir lua5.2

Thus, this commit uses the 'pcfiledir' to autodetect the path where
apt-get installed the '.pc' file.

Signed-off-by: Ricardo Ungerer <ungerer.ricardo@gmail.com>
Reviewed-by: Mark Jonas <toertel@gmail.com>