tree-wide: refactor versioning passing
The currently used mechanism is problematic for non-upstream builds.

If custom Git tags are used the version could contain information we
don't want to be actually contained or even worse - no version.

For example if one tags with "debian/2024.05+dfsg-1" the version
available in Lua would always be "0.0".

From now on the main version parts are read from the top Makefile.

Only the EXTRAVERSION is dynamically created based on the current Git
status.

In case the build is started on a tagged commit without any local
changes no EXTRAVERSION is added.

In case the build is started on a commit without any local changes the
short git SHA is added as EXTRAVERSION.

In case the build is started with local changes "-dirty" is appended to
the EXTRAVERSION.

Changes since v1:
- Generate EXTRAVERSION dynamically

Signed-off-by: Daniel Braunwarth <oss@braunwarth.dev>