mongoose: don't require /upload and /restart endpoint at root of server
When running swupdate behind a reverse proxy, swupdate might not be running at
the root / of the server but at e.g. /swupdate.  In the latter case, depending
on the configuration of the reverse proxy, the URI towards swupdate can be
either /upload or /swupdate/upload.

Currently, reverse proxy configurations where the URI /swupdate/upload gets
transformed to /upload work. However when digest auth is enabled in swupdate,
the URI becomes part of the password hash and the browser will calculate a
hash with the URI /swupdate/upload whereas swupdate will calculate a hash
with URI /upload.  These URIs are different and hence authorization will
fail.

To fix this, reverse proxy configurations should not alter the path towards
swupdate and a request from the browser to /swupdate/upload should result in a
URI in swupdate to /swupdate/upload.

However, the upload path is currently only matched on /upload and does not
match for e.g.  /swupdate/upload so uploading a file to /swupdate/upload
results in a 404 Not found.

According to mongoose mg_match documentation, "# matches zero or more
characters" so by applying this patch, file uploads will be accepted not only
at /upload but also at /swupdate/upload or even /here/is/another/path/upload.

Signed-off-by: Sam Van Den Berge <sam.van.den.berge@gmail.com>