suricatta: fix out-of-tree build
For out-of-tree builds we need to make sure to correct paths are passed
to the linker to embed Lua code.

If CONFIG_EMBEDDED_SURICATTA_LUA_SOURCE is a relative path, we imply
that the path to the Lua code to be embedded is relative to the source
directory. Otherwise absolute paths are used without any manipulation.

Signed-off-by: Daniel Braunwarth <oss@braunwarth.dev>