Revert "parser: set_find_path: properly return to-free string array"
This reverts commit cefc5ae2e43c78697281c529a29b61c4bdb3c037.

The commit causes a regression bug by following links in the
sw-description if the root of the tree is redirected via a link. An
example to create the issue is:

	stable : {
              copy1: {
                partitions: (
			..........
                );
                images: (
		        ..........
                );
                scripts: (
		        ..........
                );
              };

              copy2: {
                 ref = "#./copy1";
              };
        };

An update with -e stable,copy1 works, while -e stable,copy2 does not
find anything to be installed because the link is not followed.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reported-by: Konstantin Shabalovski <shabalovskikn@gmail.com>
Reported-by: Florian Amstutz <florian.amstutz@scs.ch>