parser: fix issues related to commit cefc5ae2
The commit above fixes a memory leak, but on a structure that was mainly
used for debugging and points to used node array in case of links. If
this structure is freed, links recursion does not work anymore.

After reverting the commit above, remove the structure causing the
regression bug because it is not really used.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>