cli: suricatta: fail on non-option argument to '-u' for server_general
The suricatta part of commit commit 14bed359464b ("cli: fail if
non-option arguments passed to sub-arguments") was not correct because
suricatta splits its parsing between start_suricatta() and each server's
server_start() functions, so the common code cannot error on leftover
non-option arguments (which actually are options' parameters)

Instead, add the check to server_general that knows what arguments to
expect (server_hawkbit already had the check, server_lua would need to
check in the user's lua code)

Note that unlike the downloader/mongoose added in previous commit, this
does not fail on unknown arguments (e.g. --foobar would be silently
ignored)
In practice, this already only works because 'S:' is declared in getopt
parameters, so we could added e/d as well and also check for invalid
arguments without too much effort if desired.
(If we did not have 'S:' in getopt arguments, we'd be left with -S'
parameter as positional argument, which I assume is why it was added in
the first place)

While here, harmonize the style of server_hawkbit's check:
- check optind < argc like the rest of the code for easier grepping
(the case compares optind >= argc with optind first so it also makes
sense from this point of view)
- use ERROR instead of fprintf like server_general

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>