cpio_utils: WARN on excessive bytes after EOF
When a cpio with additional (excessive) bytes after EOF + trailer
padding is fed into SWUpdate core, it is installed successfully.
Then, feeding the excessive bytes into SWUpdate core results in
EPIPE reported to the client as the update is (successfully) done.
The client treats this as installation error due to the return
code / errno and cannot know about the successful installation
(without, e.g., also listening on the progress interface).

Issue a warning in this case to highlight the malformed swu.

Reported-by: Wieland Eckert <wieland.eckert@siemens-healthineers.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>