suricatta/lua: Introduce custom send header configuration
Introduce a (configuration file-only) setting to add custom
headers to send to the suricatta server like the following
snippet (also) sending the custom HTTP header "X-Location":

custom-http-headers : (
    { name = "X-Location"; value="Munich,Bavaria" },
);

This is in accordance with 859be4a adding this functionality
to the hawkBit suricatta server implementation.

While at it, introduce handling a specific "wfx" configuration
file section in accordance with fc9fbe8.

Signed-off-by: Christian Storm <christian.storm@siemens.com>