suricatta: Fix out-of-tree build
With commit 16554cd, CONFIG_EMBEDDED_SURICATTA_LUA_SOURCE is prepended
by $(srctree)/ if a / is found in CONFIG_EMBEDDED_SURICATTA_LUA_SOURCE.
This is the case for absolute paths as well as for relative paths such
as the default "suricatta/server_wfx.lua" if CONFIG_SURICATTA_WFX.
While the latter case resolves correctly, the former is broken.

Hence, properly tokenize CONFIG_EMBEDDED_SURICATTA_LUA_SOURCE and
prepend $(srctree)/ only if / is not found in the first token, i.e,
it's an absolute path.
While at it, issue an error message if the relative path referenced
to by CONFIG_EMBEDDED_SURICATTA_LUA_SOURCE cannot be resolved.

Signed-off-by: Christian Storm <christian.storm@siemens.com>