Lua: Fix embedded Lua handlers out-of-tree build
Just as done for embedded Lua suricatta modules (e.g., server_wfx.lua),
fix Lua handlers embedding in out-of-tree builds by properly tokenizing
CONFIG_EMBEDDED_LUA_HANDLER_SOURCE and prepending $(srctree)/ if it is
an (assumed to be) relative path.

As the embedded Lua handler filename is an absolute path now and is
therefore no longer under SWUpdate's naming control like a relative
path would be, more characters allowed in filenames but not in a
C identifier must be handled while mangling the filename.

Signed-off-by: Christian Storm <christian.storm@siemens.com>