Makefile.flags: Fix handling of EXTRAVERSION
1. Unwanted dash in version string when EXTRAVERSION is empty
The current logic always appends a dash (-), even if EXTRAVERSION is unset.
This patch ensures the dash and EXTRAVERSION are only appended if
EXTRAVERSION is non-empty.

2. git describe fails in out-of-tree Yocto builds
Yocto uses separate source and build directories, causing git describe to
fail when invoked outside the source tree. This is fixed by running
git describe in a subshell with cd into the source directory.

Fixes: d275fec4bb2d4a0bc75f32853eca4794634d9dc4

Signed-off-by: Michael Glembotzki <Michael.Glembotzki@iris-sensing.com>