Prepare to introduce "update-type" feature
Some configurations can be grouped for update of a specific type, for
example it is plausible to have "OS update" and "application update"
(and many other) and each of them can have its set of settings.

Create a structure to group update-type configuration settings and move
them away from the main structure.

This is in preparation to introduce the "update-type" feature

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>