FEATURE: introduce update type
It is often useful to differentiate on update of different types, for
example the update for the OS can hgave a different setup and a
different acceptance of versions than an application update. This
introduces "update_type" as attribute in sw-description, and a set of
parameters can be added to swupdate.cfg. They are at the moment:
	- preupdatecmd
	- postupdatecmd
	- min and maxversion
	- no-reinstalling version

It is then possible to set a no-downgrading just for a specific type, or
for all types but with different versions.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Tested-by: Ian Dannapel <iansdannapel@gmail.com>