Prepare to use multiple crypto engines
It is current possible to select one crypto library between mbedTLS,
openSSL and WolfSSL, but they are mixing up the algorithms that can be
used. Setting pkcs#11 constraints to use WolfSSL, but it should be
possible to introduce same method to retrieve the key for other
libraries. This starts to cleanup the crypto functions in SWUpdate, and
allows to register different crypto libraries, later at the same time.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Tested-by: Michael Glembotzki <Michael.Glembotzki@iris-sensing.com>