Make SWUpdate core unaware of internal digest structures
Up now the core knows which is the internal structure used by the crypto
library to verify sw-description. The structure strongly depends on the
selected crypto library and this forbids to build SWUpdate with support
for multiple verification methods, and just one crypto library can be
linked at once. Structures are already allocated by the initialization
function for the crypto service, and it does not need to be known
outside the service, so use opaque pointer in the core.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Tested-by: Michael Glembotzki <Michael.Glembotzki@iris-sensing.com>