Add reboot to the list of parms for update type
Let that an update type can set if a reboot will be required or not,
without having to set it explicitely in sw-description. This simplifies
the setup for the user, because it is statically defined if an update of
a certain type should be followed by a reboot or not.

For compatibility, setting "reboot" in sw-description is still allowed,
and this has a higher priority than the value for the type.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Suggested-by: Christian Storm <christian.storm@siemens.com>
Suggested-by: Ian Dannapel <Ian.Dannapel@iris-sensing.com>