version: Do not print sublevel if it is zero
Prior to commit d275fec4bb2d ("tree-wide: refactor versioning passing")
the sublevel information was not printed when it did not exist.

To keep the same behavior add a check for the sublevel and only print it
if it is non-zero.

Before:

$ swupdate --version
SWUpdate v2025.05.0-1c04eb9f2ff0

After:

$ swupdate --version
SWUpdate v2025.05-1c04eb9f2ff0

Fixes: d275fec4bb2d ("tree-wide: refactor versioning passing")
Signed-off-by: Fabio Estevam <festevam@gmail.com>